Added recipes content
In this commit have added the actual recipes in teh recipe webpages.
The format for the content added is recipe name followed by a an image
of the same along with a description of the same.
Post the description follows a paragraph describing the recipe and in
following sections contains the ingriedients and steps to make the same.